+++ Proyecto/client/src/Redux/actions/shoppingBagsActions/shoppingBag_actions.js
@@ -1,6 +1,6 @@
 // import axios from 'axios';
 import Cookie from 'js-cookie';
-import { ADD_TO_CARD } from './actions_type'
+import { ADD_TO_CARD } from '../actions_type'
 //DELETE_CART, ERROR_MESSAGE, REMOVE_FROM_CART, UPDATE_FROM_CART
 
 // Crear funcion addToCart paramentros:id, qty, userID
